feat: Add ScalaNative support
